Refactoring: mappers package is created. Some classes are renamed to *Mapper
